Refactory configuração do heapsort
Refatorando classe main e configurando a implementação do heapsort

+++ MensureSortMethods/src/main/java/br/uem/paa/sortMethods/MergeSort.java
@@ -6,13 +6,13 @@ public class MergeSort implements SortMethod {
 
     static long time = 500;
     @Override
-    public long run(long[] array) {
+    public long run(int[] array) {
         time = time * 2;
         mergeSortable(array);
         return time;
     }
 
-    private long[] mergeSortable(long[] array) {
+    private int[] mergeSortable(int[] array) {
         //TODO colar aqui implementação do MergeSortFilipado.
 
         return array;
