+++ MensureSortMethods/src/main/java/br/uem/paa/util/SortMethod.java
@@ -2,6 +2,6 @@ package br.uem.paa.util;
 
 public interface SortMethod {
 
-	public long run(long[] array);
+	public long run(int[] array);
 
 }
